+++ src/bioservices/__init__.py
@@ -96,6 +96,9 @@ from .psicquic import *
 from . import ncbiblast
 from .ncbiblast import *
 
+from . import readseq
+from .readseq import *
+
 from . import biogrid
 from .biogrid import *