+++ src/bioservices/apps/fasta.py
@@ -323,7 +323,7 @@ class FASTA(object):
         print("get_fasta is deprecated. Use load_fasta instead")
         from bioservices import UniProt
         u = UniProt(verbose=False)
-        res = u.searchUniProtId(id_, frmt="fasta")
+        res = u.retrieve(id_, frmt="fasta")
         self._fasta = res[:]
         return res
 
@@ -342,7 +342,7 @@ class FASTA(object):
         from bioservices import UniProt
         u = UniProt(verbose=False)
         try:
-            res = u.searchUniProtId(id_, frmt="fasta")
+            res = u.retrieve(id_, frmt="fasta")
             # some entries in uniprot are valid but obsolet and return empty string
             if res == "":
                 raise Exception